fix dynamic shape (#890)
* fix dynamic shape

* add index_put_ shape infer
