+++ pytorch_blade/tests/quantization/test_data_prepare.py
@@ -16,6 +16,7 @@ from tests.quantization import QuantizationTestCase
 from torch import nn
 from torch.testing import FileCheck
 from torch_blade import tensorrt
+from torch_blade.config import Config
 from torch_blade.clustering.support_fusion_group import supported_node_fusion
 from torch_blade.pass_manager import _optimize_common
 from torch_blade.quantization.prepare_data import DataCollectObserver, DataPreparer
@@ -23,11 +24,14 @@ from torch_blade.tensorrt import is_available as is_tensorrt_available
 
 
 def prepare_for_data_collect(model):
-    optimized_c_module = _optimize_common(model._c)
-    model._reconstruct(optimized_c_module)
-    graph = model._c.forward.graph
-    unsupported = tensorrt.get_unsupported_nodes(graph)
-    supported_node_fusion(graph, graph, unsupported)
+    cfg = Config.get_current_context_or_new()
+    cfg.optimization_pipeline = tensorrt.backend_name()
+    with cfg:
+        optimized_c_module = _optimize_common(model._c)
+        model._reconstruct(optimized_c_module)
+        graph = model._c.forward.graph
+        unsupported = tensorrt.get_unsupported_nodes(graph)
+        supported_node_fusion(graph, graph, unsupported)
     return model
